+++ app/src/main/java/com/mexator/petfoodinspector/di/RepositoryModule.kt
@@ -4,6 +4,7 @@ import com.mexator.petfoodinspector.data.UserDataSource
 import com.mexator.petfoodinspector.data.network.PetFoodAPI
 import com.mexator.petfoodinspector.data.network.RemoteFavoritesDataSource
 import com.mexator.petfoodinspector.data.network.RemoteFoodsDataSource
+import com.mexator.petfoodinspector.data.network.RemoteUserDataSource
 import com.mexator.petfoodinspector.domain.datasource.FavouriteFoodsDataSource
 import com.mexator.petfoodinspector.domain.datasource.FoodDataSource
 import dagger.Module
@@ -15,7 +16,7 @@ class RepositoryModule {
     @Singleton
     @Provides
     fun getUserDataSource(petFoodAPI: PetFoodAPI): UserDataSource =
-        RemoteFoodsDataSource(petFoodAPI)
+        RemoteUserDataSource(petFoodAPI)
 
     @Singleton
     @Provides
